fix: revert time_remaining dtype int16 → int32 para suportar mercados 1d
int16 (max 32767) estoura para ciclos de 1d (time_remaining até 86400s) e
para janelas T_WIN de 1d (até 57600s no grid). A "otimização" de memória
quebrou o simulador Cython com 'Buffer dtype mismatch expected i32 got short'
porque o .so compilado não correspondia ao código Python.

Arquivos corrigidos:
- src/cy/sim_core.pyx: i16 → i32 para time_remaining nas 3 funções
- src/py/clean.py: pa.int16 → pa.int32 no LEAN_SCHEMA e array numpy
- src/py/clean_botlog.py: np.int16 → np.int32 no array time_remaining
- src/py/pack.py: np.int16 → np.int32 no ascontiguousarray
- src/py/gridsearch.py, src/py/validate.py: dtype int16 → int32
- run_evaluate_all_markets.py, run_validate_oos_per_market.py: idem

Após este commit: executar 'python setup.py build_ext --inplace' para
recompilar o .so com o tipo correto.

+++ run_evaluate_all_markets.py
@@ -66,7 +66,7 @@ def _eval_one_market(args: tuple) -> dict:
     data   = np.load(month_npz,  mmap_mode="r")
     cycles = np.load(cycles_npz, mmap_mode="r")
 
-    time_remaining  = np.ascontiguousarray(data["time_remaining"],    dtype=np.int16)
+    time_remaining  = np.ascontiguousarray(data["time_remaining"],    dtype=np.int32)
     prob_up         = np.ascontiguousarray(data["prob_up"],           dtype=np.float32)
     cycle_start_idx = np.ascontiguousarray(cycles["cycle_start_idx"], dtype=np.int32)
     cycle_end_idx   = np.ascontiguousarray(cycles["cycle_end_idx"],   dtype=np.int32)
